+++ src/t_string.c
@@ -256,11 +256,11 @@ void getrangeCommand(redisClient *c) {
     }
 
     /* Convert negative indexes */
-    if (start < 0) start = strlen+start;
-    if (end < 0) end = strlen+end;
+    if (start < 0) start = (long)(strlen+start);
+    if (end < 0) end = (long)(strlen+end);
     if (start < 0) start = 0;
     if (end < 0) end = 0;
-    if ((unsigned)end >= strlen) end = strlen-1;
+    if ((unsigned)end >= strlen) end = (long)(strlen-1);
 
     /* Precondition: end >= 0 && end < strlen, so the only condition where
      * nothing can be returned is: start > end. */
